fix: remove protocolFee refunds
from 0x's Discord:
"The protocol fee was a fee paid by takers and ultimately rebated to market makers when their orders were filled. However, the community voted to remove protocol fees in ZEIP-91, which decreased the protocol fee multiplier from the current value (70,000) to zero (0) for v3 onward. So, there is no longer a protocol fee. The fee was paid in ETH."